@@ -2,11 +2,11 @@ package com.javatechie.jwt.api;
 
 import com.javatechie.jwt.api.entity.User;
 import com.javatechie.jwt.api.repository.UserRepository;
+import jakarta.annotation.PostConstruct;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-import javax.annotation.PostConstruct;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
